Delete User Completed

+++ app/Http/Controllers/Admin/UserManagermentController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
 use App\User;
+use Auth;
 
 class UserManagermentController extends Controller
 {
@@ -26,13 +27,20 @@ class UserManagermentController extends Controller
         ]);
     }
 
-    public function postDeleteUser($id){
+    public function getDeleteUser($id){
         $deleteUser = User::findOrFail($id);
-        $nameDeleteUser = $deleteUser->username;
-        $deleteUser->delete();
-        return redirect()->route('admin.dashboard.getUser')->with([
-            'msgAlert' => 'Bạn đã xóa thành công user : ' . $nameDeleteUser,
-            'lvlAlert' => 'success'
-        ]);
+        if ($id === Auth::id() || Auth::user()->level !== 1 || $deleteUser->level === 1){
+            return redirect()->route('admin.dashboard.getUser')->with([
+                'msgAlert' => 'Bạn không có quyền xóa username này ! ',
+                'lvlAlert' => 'danger'
+            ]);
+        }else{
+            $nameDeleteUser = $deleteUser->username;
+            $deleteUser->delete();
+            return redirect()->route('admin.dashboard.getUser')->with([
+                'msgAlert' => 'Bạn đã xóa thành công username : ' . $nameDeleteUser,
+                'lvlAlert' => 'success'
+            ]);
+        }
     }
 }
